+++ src/scriptlets/set-cookie.js
@@ -45,6 +45,7 @@ import {
  *         - `disable` / `disabled`
  *         - `necessary` / `required`
  *         - `hide` / `hidden`
+ *         - `essential` / `nonessential`
  * - `path` — optional, cookie path, defaults to `/`; possible values:
  *     - `/` — root path
  *     - `none` — to set no path at all